Add mixin class for unique distribution
